% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisLabels.R
\name{amAxisBreaks}
\alias{amAxisBreaks}
\title{Axis breaks}
\usage{
amAxisBreaks(
  values = NULL,
  labels = NULL,
  interval = NULL,
  timeInterval = NULL
)
}
\arguments{
\item{values}{positions of the breaks, a vector of values; for a date axis,
this must be a vector of dates}

\item{labels}{if \code{values} is given, the labels of the breaks; if
\code{NULL}, the labels are set to the values}

\item{interval}{for equally spaced breaks, the number of pixels between two
consecutive breaks; ignored if \code{values} is given}

\item{timeInterval}{for equally spaced breaks on a date axis, this option
defines the interval between two consecutive breaks; it must be a string
like \code{"1 day"}, \code{"7 days"}, \code{"1 week"}, \code{"2 months"},
...; ignored if \code{values} or \code{interval} is given}
}
\description{
Create an object defining the breaks on an axis.
}
