% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisLabels.R
\name{amAxisLabels}
\alias{amAxisLabels}
\title{Axis labels}
\usage{
amAxisLabels(
  color = NULL,
  fontSize = 18,
  rotation = 0,
  formatter = NULL,
  timeInterval = NULL
)
}
\arguments{
\item{color}{color of the labels}

\item{fontSize}{font size of the labels}

\item{rotation}{rotation angle}

\item{formatter}{this option defines the format of the axis labels;
this should be a
\href{https://www.amcharts.com/docs/v4/concepts/formatters/formatting-numbers/}{number formatting string}
for a numeric axis, and a list created with
\code{\link{amDateAxisFormatter}} for a date axis}

\item{timeInterval}{for a date axis, this option defines the interval
between two consecutive labels; it must be a string like \code{"1 day"},
\code{"7 days"}, \code{"1 week"}, \code{"2 months"}, ...}
}
\value{
A list of settings for the labels of an axis.
}
\description{
Create a list of settings for the labels of an axis.
}
