% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amBarChart.R
\name{amBarChart-shiny}
\alias{amBarChart-shiny}
\alias{amBarChartOutput}
\alias{renderAmBarChart}
\title{Shiny bindings for \code{amBarChart}}
\usage{
amBarChartOutput(outputId, width = "100\%", height = "400px")

renderAmBarChart(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended}

\item{expr}{an expression that generates a bar chart with
\code{\link{amBarChart}}}

\item{env}{the environment in which to evaluate \code{expr}}

\item{quoted}{whether \code{expr} is a quoted expression}
}
\description{
Output and render functions for using \code{\link{amBarChart}}
within Shiny applications and interactive Rmd documents.
}
