% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{amButton}
\alias{amButton}
\title{Button}
\usage{
amButton(label, color = NULL, position = 0.9, marginRight = 10)
}
\arguments{
\item{label}{label of the button, a character string or a list created with
\code{\link{amText}} for a formatted label}

\item{color}{button color}

\item{position}{the vertical position of the button: \code{0} for bottom,
\code{1} for top}

\item{marginRight}{right margin in pixels}
}
\value{
A list of settings for a button.
}
\description{
Create a list of settings for a button.
}
