% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{amc_yAxes}
\alias{amc_yAxes}
\title{Add yAxes parameters}
\usage{
amc_yAxes(x, ...)
}
\arguments{
\item{x}{A \code{rAmCharts4} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{https://www.amcharts.com/docs/v4/}.}
}
\value{
A \code{rAmCharts4} \code{htmlwidget} object.
}
\description{
Add yAxes parameters
}
