% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnStyle.R
\name{amColumn}
\alias{amColumn}
\title{Columns style}
\usage{
amColumn(color = NULL, strokeColor = NULL, strokeWidth = 4, cornerRadius = 8)
}
\arguments{
\item{color}{color of the columns}

\item{strokeColor}{color of the border of the columns}

\item{strokeWidth}{width of the border of the the columns}

\item{cornerRadius}{radius of the corners of the columns}
}
\value{
A list of settings for usage in \code{\link{amBarChart}}
  or \code{\link{amHorizontalBarChart}}
}
\description{
Create a list of settings for the columns of a bar chart.
}
