% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{amFont}
\alias{amFont}
\title{Font}
\usage{
amFont(fontSize = NULL, fontWeight = "normal", fontFamily = NULL)
}
\arguments{
\item{fontSize}{font size, must be given as a character string like
\code{"10px"} or \code{"2em"}, or a numeric value, the font size in pixels}

\item{fontWeight}{font weight, it can be \code{"normal"},
\code{"bold"}, \code{"bolder"}, \code{"lighter"}, or a number in
\code{seq(100, 900, by = 100)}}

\item{fontFamily}{font family}
}
\value{
A list of settings for a font.
}
\description{
Create a list of settings for a font.
}
\note{
There is no option for the font style.
}
