% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand.R
\name{amHand}
\alias{amHand}
\title{Gauge hand}
\usage{
amHand(innerRadius, width, color, strokeColor)
}
\arguments{
\item{innerRadius}{inner radius of the hand, given as a percentage}

\item{width}{width of the base of the hand in pixels, a positive number}

\item{color}{color of the hand}

\item{strokeColor}{stroke color of the hand}
}
\value{
A list of settings for the hand of a gauge chart.
}
\description{
Create a list of settings for the hand of a gauge chart.
}
