% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{amImage}
\alias{amImage}
\title{Image}
\usage{
amImage(href, width, height, opacity = 1)
}
\arguments{
\item{href}{a link to an image file or a base64 string representing an
image (you can create such a string with \code{base64enc::dataURI})}

\item{width, height}{dimensions of the image}

\item{opacity}{opacity of the image, a number between 0 and 1}
}
\value{
A list of settings for an image.
}
\description{
Create a list of settings for an image.
}
