% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{amLegend}
\alias{amLegend}
\title{Legend}
\usage{
amLegend(position = "bottom", itemsWidth = 20, itemsHeight = 20)
}
\arguments{
\item{position}{legend position}

\item{itemsWidth}{width of the legend items}

\item{itemsHeight}{height of the legend items}
}
\value{
A list of settings for a legend.
}
\description{
Create a list of settings for a legend.
}
