% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineStyle.R
\name{amLine}
\alias{amLine}
\title{Line style}
\usage{
amLine(color = NULL, width = 3, dash = NULL, tensionX = NULL, tensionY = NULL)
}
\arguments{
\item{color}{line color}

\item{width}{line width}

\item{dash}{string defining a dashed/dotted line; see
\url{https://www.amcharts.com/docs/v4/tutorials/dotted-and-dashed-lines/#Dasharray_values}}

\item{tensionX, tensionY}{parameters for the smoothing; see
\url{https://www.amcharts.com/docs/v4/chart-types/xy-chart/#Smoothed_lines}
for the meaning of these parameters}
}
\value{
A list of settings for a line.
}
\description{
Create a list of settings for a line.
}
