% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amScatterChart.R
\name{amScatterChart}
\alias{amScatterChart}
\title{Create a HTML widget displaying a line chart}
\usage{
amScatterChart(
  data,
  data2 = NULL,
  xValue,
  yValues,
  yValueNames = NULL,
  xLimits = NULL,
  yLimits = NULL,
  valueFormatter = "#.",
  chartTitle = NULL,
  theme = NULL,
  draggable = FALSE,
  tooltip = NULL,
  pointsStyle = NULL,
  backgroundColor = NULL,
  xAxis = NULL,
  yAxis = NULL,
  scrollbarX = FALSE,
  scrollbarY = FALSE,
  gridLines = NULL,
  legend = NULL,
  caption = NULL,
  button = NULL,
  width = NULL,
  height = NULL,
  chartId = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{a dataframe}

\item{data2}{\code{NULL} or a dataframe used to update the data with the
button; its column names must include the column names of \code{data}
given in \code{values} and it must have the same number of rows as
\code{data}}

\item{xValue}{name of the column of \code{data} to be used on the
x-axis}

\item{yValues}{name(s) of the column(s) of \code{data} to be used on the
y-axis}

\item{yValueNames}{names of the variables on the y-axis,
to appear in the legend;
\code{NULL} to use \code{yValues} as names, otherwise a named list of the
form \code{list(yvalue1 = "ValueName1", yvalue2 = "ValueName2", ...)} where
\code{yvalue1}, \code{yvalue2}, ... are the column names given in
\code{yValues} and \code{"ValueName1"}, \code{"ValueName2"}, ... are the
desired names to appear in the legend}

\item{xLimits}{range of the x-axis, a vector of two values specifying
the left and the right limits of the x-axis; \code{NULL} for default values}

\item{yLimits}{range of the y-axis, a vector of two values specifying
the lower and the upper limits of the y-axis; \code{NULL} for default values}

\item{valueFormatter}{a number formatter; see
\url{https://www.amcharts.com/docs/v4/concepts/formatters/formatting-numbers/}}

\item{chartTitle}{chart title, \code{NULL}, character, or list of settings}

\item{theme}{theme, \code{NULL} or one of \code{"dataviz"},
\code{"material"}, \code{"kelly"}, \code{"dark"}, \code{"moonrisekingdom"},
\code{"frozen"}, \code{"spiritedaway"}, \code{"patterns"},
\code{"microchart"}}

\item{draggable}{\code{TRUE}/\code{FALSE} to enable/disable dragging of
all lines, otherwise a named list of the form
\code{list(yvalue1 = TRUE, yvalue2 = FALSE, ...)} to enable/disable the
dragging for each bar corresponding to a column given in \code{yValues}}

\item{tooltip}{tooltip settings given as a list, or just a string for the
\code{text} field, or \code{FALSE} for no tooltip, or \code{NULL} for the
default tooltip; the \code{text} field must be a formatted string:
\url{https://www.amcharts.com/docs/v4/concepts/formatters/formatting-strings/}}

\item{pointsStyle}{settings of the points style; \code{NULL} for default,
otherwise a named list of the form
\code{list(yvalue1 = settings1, yvalue2 = settings2, ...)} where
\code{settings1}, \code{settings2}, ... are lists created with
\code{\link{amCircle}}, \code{\link{amTriangle}} or
\code{\link{amRectangle}}; this can also be a
single list of settings that will be applied to each series}

\item{backgroundColor}{a color for the chart background}

\item{xAxis}{settings of the x-axis given as a list, or just a string
for the axis title}

\item{yAxis}{settings of the y-axis given as a list, or just a string
for the axis title}

\item{scrollbarX}{logical, whether to add a scrollbar for the category axis}

\item{scrollbarY}{logical, whether to add a scrollbar for the value axis}

\item{gridLines}{settings of the grid lines}

\item{legend}{logical, whether to display the legend}

\item{caption}{settings of the caption, or \code{NULL} for no caption}

\item{button}{\code{NULL} for the default, \code{FALSE} for no button,
a single character string giving the button label,
or settings of the button given as
a list with these fields: \code{text} for the button label, \code{color} for
the label color, \code{fill} for the button color, and \code{position}
for the button position as a percentage (\code{0} for bottom,
\code{1} for top); this button is used to replace the current data
with \code{data2}}

\item{width}{the width of the chart, e.g. \code{"600px"} or \code{"80\%"};
ignored if the chart is displayed in Shiny, in which case the width is
given in \code{\link{amChart4Output}}}

\item{height}{the height of the chart, e.g. \code{"400px"};
ignored if the chart is displayed in Shiny, in which case the width is
given in \code{\link{amChart4Output}}}

\item{chartId}{a HTML id for the chart}

\item{elementId}{a HTML id for the container of the chart; ignored if the
chart is displayed in Shiny, in which case the id is given by the Shiny id}
}
\description{
Create a HTML widget displaying a line chart.
}
\note{
A color can be given by the name of a R color, the name of a CSS
color, e.g. \code{"transparent"} or \code{"fuchsia"}, an HEX code like
\code{"#ff009a"}, a RGB code like \code{"rgb(255,100,39)"}, or a HSL code
like \code{"hsl(360,11,255)"}.
}
\examples{
# iris data: petal widths ####
dat <- iris
dat$obs <- rep(1:50, 3)
dat <- reshape2::dcast(dat, obs ~ Species, value.var = "Petal.Width")

amScatterChart(
  data = dat,
  width = "700px",
  xValue = "obs",
  yValues = c("setosa", "versicolor", "virginica"),
  draggable = FALSE,
  backgroundColor = "#30303d",
  pointsStyle = list(
    color = list(
      setosa = "orange", versicolor = "cyan", virginica = "palegreen"
     ),
    strokeColor = list(
      setosa = "red", versicolor = "blue", virginica = "green"
     ),
    strokeWidth = 2
  ),
  tooltip = "obs: {valueX}\nvalue: {valueY}",
  chartTitle = list(text = "Iris data", color = "whitesmoke"),
  xAxis = list(title = list(text = "Observation",
                            fontSize = 21,
                            color = "silver"),
               labels = list(color = "whitesmoke",
                             fontSize = 17)),
  yAxis = list(title = list(text = "Petal width",
                            fontSize = 21,
                            color = "silver"),
               labels = list(color = "whitesmoke",
                             fontSize = 14)),
  valueFormatter = "#.#",
  caption = list(text = "[font-style:italic]rAmCharts4[/]",
                 color = "yellow"),
  gridLines = list(color = "whitesmoke",
                   opacity = 0.4,
                   width = 1),
  theme = "dark")


# iris data: petal widths vs petal lengths

dat <- iris
dat$obs <- rep(1:50, 3)
dat <-
  reshape2::dcast(dat, obs + Petal.Length ~ Species, value.var = "Petal.Width")

amScatterChart(
  data = dat,
  width = "700px",
  xValue = "Petal.Length",
  yValues = c("setosa", "versicolor", "virginica"),
  draggable = FALSE,
  backgroundColor = "#30303d",
  pointsStyle = list(
    setosa = amCircle(color = "orange", strokeColor = "red"),
    versicolor = amCircle(color = "cyan", strokeColor = "blue"),
    virginica = amCircle(color = "palegreen", strokeColor = "darkgreen")
  ),
  tooltip = "length: {valueX}\nwidth: {valueY}",
  chartTitle = list(text = "Iris data", color = "silver"),
  xAxis = list(title = list(text = "Petal length",
                            fontSize = 19,
                            color = "gold"),
               labels = list(color = "whitesmoke",
                             fontSize = 17)),
  yAxis = list(title = list(text = "Petal width",
                            fontSize = 19,
                            color = "gold"),
               labels = list(color = "whitesmoke",
                             fontSize = 17)),
  valueFormatter = "#.#",
  caption = list(text = "[font-style:italic]rAmCharts4[/]",
                 color = "yellow"),
  gridLines = list(color = "whitesmoke",
                   opacity = 0.4,
                   width = 1),
  theme = "dark")
}
