% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentStyle.R
\name{amSegment}
\alias{amSegment}
\title{Segment style}
\usage{
amSegment(color = NULL, width = 1)
}
\arguments{
\item{color}{color of the segment; this can be a
\link[rAmCharts4:amColorAdapterFromVector]{color adapter}}

\item{width}{width of the segment}
}
\value{
A list of settings for a segment.
}
\description{
Create a list of settings for a segment.
}
\note{
A color can be given by the name of a R color, the name of a CSS
  color, e.g. \code{"lime"} or \code{"indigo"}, an HEX code like
  \code{"#ff009a"}, a RGB code like \code{"rgb(255,100,39)"}, or a HSL code
  like \code{"hsl(360,11,255)"}.
}
