% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amCharts.R
\name{amWordCloud}
\alias{amWordCloud}
\title{Word Cloud}
\usage{
amWordCloud(..., width = NULL, height = NULL)
}
\description{
Word Cloud
}
\examples{
\dontrun{
amWordCloud(data = data.frame(tag = c("My", "Best", "Word"),
                              weight = c(100, 200, 300)),
            series = list(list(type = "WordCloudSeries",
                               dataFields = dataFields(
                                 word = "tag",
                                 value = "weight"))))

}

}
