% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoomButtons.R
\name{amZoomButtons}
\alias{amZoomButtons}
\title{Zoom buttons}
\usage{
amZoomButtons(
  halign = "left",
  valign = "top",
  marginH = 5,
  marginV = 5,
  zoomFactor = 0.1
)
}
\arguments{
\item{halign}{\code{"left"} or \code{"right"}}

\item{valign}{\code{"top"} or \code{"bottom"}}

\item{marginH}{horizontal margin}

\item{marginV}{vertical margin}

\item{zoomFactor}{zoom factor}
}
\value{
A list of parameters for zoom buttons, for usage in
  \code{\link{amLineChart}} or \code{\link{amScatterChart}}
}
\description{
Zoom buttons.
}
