% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{dataFields}
\alias{dataFields}
\title{Add dataFields information}
\usage{
dataFields(value = NULL, category = NULL, valueY = NULL,
  categoryX = NULL, ...)
}
\arguments{
\item{value}{A \code{character} name of value field.}

\item{category}{A \code{character} name of category field.}

\item{valueY}{A \code{character} name of Y field value.}

\item{categoryX}{A \code{character} name of categoryX field.}

\item{...}{Arguments defined in \url{https://www.amcharts.com/docs/v4/}.}
}
\value{
A \code{list}.
}
\description{
Add dataFields information
}
