% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{rAmCharts4-adapters}
\alias{rAmCharts4-adapters}
\alias{amColorAdapterFromVector}
\alias{amColorAdapterFromCuts}
\title{Adapters}
\usage{
amColorAdapterFromVector(colors)

amColorAdapterFromCuts(cuts, colors, axis)
}
\arguments{
\item{colors}{vector of colors}

\item{cuts}{a vector of cut points (sorted increasingly)}

\item{axis}{\code{"X"} or \code{"Y"}}
}
\description{
Adapters
}
