% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rAmCharts4-import.R
\docType{import}
\name{rAmCharts4-imports}
\alias{rAmCharts4-imports}
\alias{JS}
\alias{saveWidget}
\title{Objects imported from other packages}
\description{
These objects are imported from other packages.
  Follow the links to their documentation:
  \code{\link[htmlwidgets:JS]{JS}},
  \code{\link[htmlwidgets:saveWidget]{saveWidget}}
}
