% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{rAmCharts4-shiny}
\alias{rAmCharts4-shiny}
\alias{amChart4Output}
\alias{renderAmChart4}
\title{Shiny bindings for using \code{rAmCharts4} in Shiny}
\usage{
amChart4Output(outputId, width = "100\%", height = "400px")

renderAmChart4(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended}

\item{expr}{an expression that generates a chart with
\code{\link{amBarChart}}, \code{\link{amHorizontalBarChart}} or
\code{\link{amLineChart}}}

\item{env}{the environment in which to evaluate \code{expr}}

\item{quoted}{whether \code{expr} is a quoted expression}
}
\description{
Output and render functions for using
the \code{rAmCharts4} widgets
within Shiny applications and interactive Rmd documents.
}
