% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{series}
\alias{series}
\title{Add series information}
\usage{
series(type = NULL, stroke = NULL, strokeWidth = NULL, ...)
}
\arguments{
\item{type}{A \code{character} type of serie.}

\item{stroke}{A \code{character} color, hexa.}

\item{strokeWidth}{A \code{interger} size.}

\item{...}{Arguments defined in \url{https://www.amcharts.com/docs/v4/}.}
}
\value{
A \code{list}.
}
\description{
Add series information
}
\examples{
\dontrun{
amCharts4(data = data.frame(tag = c("My", "Best", "Word"),
                            weight = c(100, 200, 300)),
          series = series(type = "WordCloudSeries",
                          dataFields = dataFields(
                            word = "tag",
                            value = "weight")),
          type = "WordCloud")
}

}
