% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyIcon.R
\name{tinyIcon}
\alias{tinyIcon}
\alias{tinyIcons}
\alias{shinyAppTinyIcons}
\title{Icons}
\usage{
tinyIcon(icon)

tinyIcons()

shinyAppTinyIcons()
}
\arguments{
\item{icon}{name of an icon; \code{tinyIcons()} returns the list of
available icons, and \code{shinyAppTinyIcons()} runs a Shiny app which
displays the available icons}
}
\value{
A base64 string that can be used in the \code{href} argument of
  \code{\link{amImage}}.
}
\description{
Icons for usage in \code{\link{amImage}}.
}
