% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{updateAmPieChart}
\alias{updateAmPieChart}
\title{Update the data of a pie chart}
\usage{
updateAmPieChart(session, outputId, data)
}
\arguments{
\item{session}{the Shiny \code{session} object}

\item{outputId}{the output id passed on to \code{\link{amChart4Output}}}

\item{data}{new data; if it is not valid, then nothing will happen (in order
to be valid it must have the same structure as the data passed on to
\code{\link{amPieChart}}); in this case check the JavaScript console, it
will report the encountered issue}
}
\description{
Update the data of a pie chart in a Shiny app.
}
