% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisLabels.R
\name{amAxisLabels}
\alias{amAxisLabels}
\title{Axis labels}
\usage{
amAxisLabels(color = NULL, fontSize = 18, rotation = 0, formatter = NULL)
}
\arguments{
\item{color}{color of the labels}

\item{fontSize}{font size of the labels}

\item{rotation}{rotation angle}

\item{formatter}{a formatting string for the labels; see
\href{https://www.amcharts.com/docs/v4/concepts/formatters/formatting-numbers/}{Formatting Numbers}
and
\href{https://www.amcharts.com/docs/v4/concepts/formatters/formatting-date-time/}{Formatting Date and Time}}
}
\value{
A list of settings for the labels of an axis.
}
\description{
Create a list of settings for the labels of an axis.
}
