% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amBarChart.R
\name{amBarChart}
\alias{amBarChart}
\title{Create a HTML widget displaying a bar chart}
\usage{
amBarChart(
  data,
  data2 = NULL,
  category,
  values,
  valueNames = NULL,
  minValue,
  maxValue,
  valueFormatter = "#.",
  chartTitle = NULL,
  theme = NULL,
  draggable = FALSE,
  tooltip = NULL,
  columnStyle = NULL,
  backgroundColor = NULL,
  cellWidth = NULL,
  columnWidth = NULL,
  xAxis = NULL,
  yAxis = NULL,
  scrollbarX = FALSE,
  scrollbarY = FALSE,
  gridLines = NULL,
  legend = NULL,
  caption = NULL,
  button = NULL,
  width = NULL,
  height = NULL,
  chartId = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{a dataframe}

\item{data2}{a dataframe used to update the data with the button}

\item{category}{name of the column of \code{data} to be used on the
category axis}

\item{values}{name(s) of the column(s) of \code{data} to be used on the
value axis}

\item{valueNames}{names of the values variables, to appear in the legend;
\code{NULL} to use \code{values} as names, otherwise a named list of the
form \code{list(value1 = "ValueName1", value2 = "ValueName2", ...)} where
\code{value1}, \code{value2}, ... are the column names given in
\code{values} and \code{"ValueName1"}, \code{"ValueName2"}, ... are the
desired names to appear in the legend}

\item{minValue}{minimum value of the y-axis}

\item{maxValue}{maximum value of the y-axis}

\item{valueFormatter}{a number formatter; see
\url{https://www.amcharts.com/docs/v4/concepts/formatters/formatting-numbers/}}

\item{chartTitle}{chart title, \code{NULL}, character, or list of settings}

\item{theme}{theme, \code{NULL} or one of \code{"dataviz"},
\code{"material"}, \code{"kelly"}, \code{"dark"}, \code{"moonrisekingdom"},
\code{"frozen"}, \code{"spiritedaway"}, \code{"patterns"},
\code{"microchart"}}

\item{draggable}{\code{TRUE}/\code{FALSE} to enable/disable dragging of
all bars, otherwise a named list of the form
\code{list(value1 = TRUE, value2 = FALSE, ...)} to enable/disable the
dragging for each bar corresponding to a column given in \code{values}}

\item{tooltip}{tooltip settings given as a list, or just a string for the
\code{text} field, or \code{NULL} for no tooltip; the \code{text} field is
given as a formatted string; see
\url{https://www.amcharts.com/docs/v4/concepts/formatters/formatting-strings/}}

\item{columnStyle}{settings of the columns style; \code{NULL} for default,
otherwise a named list with three fields: \code{fill} to set the colors
of the columns, given as a single color or a named list of the form
\code{list(value1 = "red", value2 = "green", ...)}, \code{stroke} to set
the color of the borders of the columns, and \code{cornerRadius} to set
the radius of the corners of the columns}

\item{backgroundColor}{a color for the chart background}

\item{cellWidth}{cell width in percent; for a simple bar chart, this is the
width of the columns; for a grouped bar chart, this is the width of the
clusters of columns}

\item{columnWidth}{column width, a percentage of the cell width; set to 100
for a simple bar chart and use \code{cellWidth} to control the width of the
columns; for a grouped bar chart, this controls the spacing between the
columns within a cluster of columns}

\item{xAxis}{settings of the category axis given as a list, or just a string
for the axis title}

\item{yAxis}{settings of the value axis given as a list, or just a string
for the axis title}

\item{scrollbarX}{logical, whether to add a scrollbar for the category axis}

\item{scrollbarY}{logical, whether to add a scrollbar for the value axis}

\item{gridLines}{settings of the grid lines}

\item{legend}{logical, whether to display the legend}

\item{caption}{settings of the caption, or \code{NULL} for no caption}

\item{button}{\code{NULL} for the default, \code{FALSE} for no button,
a single character string giving the button label,
or settings of the button given as
a list with these fields: \code{text} for the button label, \code{color} for
the label color, \code{fill} for the button color, and \code{position}
for the button position as a percentage (\code{0} for bottom,
\code{1} for top); this button is used to replace the current data
with \code{data2}}

\item{width}{the width of the chart, e.g. \code{"600px"} or \code{"80\%"};
ignored if the chart is displayed in Shiny, in which case the width is
given in \code{\link{amBarChartOutput}}}

\item{height}{the height of the chart, e.g. \code{"400px"};
ignored if the chart is displayed in Shiny, in which case the width is
given in \code{\link{amBarChartOutput}}}

\item{chartId}{a HTML id for the chart}

\item{elementId}{a HTML id for the container of the chart; ignored if the
chart is displayed in Shiny, in which case the id is given by the Shiny id}
}
\description{
Create a HTML widget displaying a bar chart.
}
\examples{
# a simple bar chart ####

dat <- data.frame(
  country = c("USA", "China", "Japan", "Germany", "UK", "France"),
  visits = c(3025, 1882, 1809, 1322, 1122, 1114)
)

amBarChart(
  data = dat, data2 = dat,
  width = "600px",
  category = "country", values = "visits",
  draggable = TRUE,
  tooltip = "[font-style:italic;#ffff00]{valueY}[/]",
  chartTitle =
    list(text = "Visits per country", fontSize = 22, color = "orangered"),
  xAxis = list(title = list(text = "Country", color = "maroon")),
  yAxis = list(title = list(text = "Visits", color = "maroon")),
  minValue = 0, maxValue = 4000,
  valueFormatter = "#.",
  caption = list(text = "Year 2018", color = "red"),
  theme = "material")

# a grouped barchart

set.seed(666)
dat <- data.frame(
  country = c("USA", "China", "Japan", "Germany", "UK", "France"),
  visits = c(3025, 1882, 1809, 1322, 1122, 1114),
  income = rpois(6, 25),
  expenses = rpois(6, 20)
)

amBarChart(
  data = dat,
  width = "700px",
  category = "country",
  values = c("income", "expenses"),
  valueNames = list(income = "Income", expenses = "Expenses"),
  draggable = list(income = TRUE, expenses = FALSE),
  backgroundColor = "#30303d",
  columnStyle = list(
    fill = list(income = "darkmagenta", expenses = "darkred"),
    stroke = "#cccccc"
  ),
  chartTitle = list(text = "Income and expenses per country"),
  xAxis = list(title = list(text = "Country")),
  yAxis = list(title = list(text = "Income and expenses")),
  minValue = 0, maxValue = 41,
  valueFormatter = "#.#",
  caption = list(text = "Year 2018"),
  theme = "dark")
}
