% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{amc_series}
\alias{amc_series}
\title{Add series parameters}
\usage{
amc_series(x, ...)
}
\arguments{
\item{x}{A \code{rAmCharts4} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{https://www.amcharts.com/docs/v4/}.}
}
\value{
A \code{rAmCharts4} \code{htmlwidget} object.
}
\description{
Add series parameters
}
