% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnStyle.R
\name{amColumn}
\alias{amColumn}
\title{Columns style}
\usage{
amColumn(
  color = NULL,
  opacity = NULL,
  strokeColor = NULL,
  strokeWidth = 4,
  cornerRadius = 8
)
}
\arguments{
\item{color}{color of the columns; this can be a
\link[rAmCharts4:amColorAdapterFromVector]{color adapter}}

\item{opacity}{opacity of the columns, a number between 0 and 1}

\item{strokeColor}{color of the border of the columns; this can be a
\link[rAmCharts4:amColorAdapterFromVector]{color adapter}}

\item{strokeWidth}{width of the border of the columns}

\item{cornerRadius}{radius of the corners of the columns}
}
\value{
A list of settings for usage in \code{\link{amBarChart}}
  or \code{\link{amHorizontalBarChart}}
}
\description{
Create a list of settings for the columns of a bar chart.
}
\note{
A color can be given by the name of a R color, the name of a CSS
  color, e.g. \code{"transparent"} or \code{"fuchsia"}, an HEX code like
  \code{"#ff009a"}, a RGB code like \code{"rgb(255,100,39)"}, or a HSL code
  like \code{"hsl(360,11,255)"}.
}
