% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amCharts.R
\name{amGaugeChart}
\alias{amGaugeChart}
\title{Gauge chart}
\usage{
amGaugeChart(..., width = NULL, height = NULL)
}
\description{
Gauge chart
}
\examples{
\dontrun{
   amGaugeChart(xAxes = list(list(type = "ValueAxis", min = 0, max = 100)))
}

}
