% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amLineChart.R
\name{amLineChart}
\alias{amLineChart}
\title{Create a HTML widget displaying a line chart}
\usage{
amLineChart(
  data,
  data2 = NULL,
  xValue,
  yValues,
  yValueNames = NULL,
  minY,
  maxY,
  valueFormatter = "#.",
  chartTitle = NULL,
  theme = NULL,
  draggable = FALSE,
  tooltip = NULL,
  lineStyle = NULL,
  backgroundColor = NULL,
  xAxis = NULL,
  yAxis = NULL,
  scrollbarX = FALSE,
  scrollbarY = FALSE,
  gridLines = NULL,
  legend = NULL,
  caption = NULL,
  button = NULL,
  width = NULL,
  height = NULL,
  chartId = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{a dataframe}

\item{data2}{\code{NULL} or a dataframe used to update the data with the
button; its column names must include the column names of \code{data}
given in \code{values} and it must have the same number of rows as
\code{data}}

\item{xValue}{name of the column of \code{data} to be used on the
x-axis}

\item{yValues}{name(s) of the column(s) of \code{data} to be used on the
y-axis}

\item{yValueNames}{names of the variables on the y-axis,
to appear in the legend;
\code{NULL} to use \code{yValues} as names, otherwise a named list of the
form \code{list(yvalue1 = "ValueName1", yvalue2 = "ValueName2", ...)} where
\code{yvalue1}, \code{yvalue2}, ... are the column names given in
\code{yValues} and \code{"ValueName1"}, \code{"ValueName2"}, ... are the
desired names to appear in the legend}

\item{minY}{minimum value of the y-axis}

\item{maxY}{maximum value of the y-axis}

\item{valueFormatter}{a number formatter; see
\url{https://www.amcharts.com/docs/v4/concepts/formatters/formatting-numbers/}}

\item{chartTitle}{chart title, \code{NULL}, character, or list of settings}

\item{theme}{theme, \code{NULL} or one of \code{"dataviz"},
\code{"material"}, \code{"kelly"}, \code{"dark"}, \code{"moonrisekingdom"},
\code{"frozen"}, \code{"spiritedaway"}, \code{"patterns"},
\code{"microchart"}}

\item{draggable}{\code{TRUE}/\code{FALSE} to enable/disable dragging of
all lines, otherwise a named list of the form
\code{list(yvalue1 = TRUE, yvalue2 = FALSE, ...)} to enable/disable the
dragging for each bar corresponding to a column given in \code{yValues}}

\item{tooltip}{tooltip settings given as a list, or just a string for the
\code{text} field, or \code{FALSE} for no tooltip, or \code{NULL} for the
default tooltip; the \code{text} field must be a formatted string:
\url{https://www.amcharts.com/docs/v4/concepts/formatters/formatting-strings/}}

\item{lineStyle}{settings of the lines style; \code{NULL} for default,
otherwise a named list with XXXX three fields: \code{color} to set the colors
of the lines, given as a single color or a named list of the form
\code{list(yvalue1 = "red", yvalue2 = "green", ...)}, \code{width} to set
the width of the lines, given as a single number or a named list of the form
\code{list(yvalue1 = 1, yvalue2 = 3, ...)}}

\item{backgroundColor}{a color for the chart background}

\item{xAxis}{settings of the x-axis given as a list, or just a string
for the axis title}

\item{yAxis}{settings of the y-axis given as a list, or just a string
for the axis title}

\item{scrollbarX}{logical, whether to add a scrollbar for the category axis}

\item{scrollbarY}{logical, whether to add a scrollbar for the value axis}

\item{gridLines}{settings of the grid lines}

\item{legend}{logical, whether to display the legend}

\item{caption}{settings of the caption, or \code{NULL} for no caption}

\item{button}{\code{NULL} for the default, \code{FALSE} for no button,
a single character string giving the button label,
or settings of the button given as
a list with these fields: \code{text} for the button label, \code{color} for
the label color, \code{fill} for the button color, and \code{position}
for the button position as a percentage (\code{0} for bottom,
\code{1} for top); this button is used to replace the current data
with \code{data2}}

\item{width}{the width of the chart, e.g. \code{"600px"} or \code{"80\%"};
ignored if the chart is displayed in Shiny, in which case the width is
given in \code{\link{amChart4Output}}}

\item{height}{the height of the chart, e.g. \code{"400px"};
ignored if the chart is displayed in Shiny, in which case the width is
given in \code{\link{amChart4Output}}}

\item{chartId}{a HTML id for the chart}

\item{elementId}{a HTML id for the container of the chart; ignored if the
chart is displayed in Shiny, in which case the id is given by the Shiny id}
}
\description{
Create a HTML widget displaying a line chart.
}
\note{
A color can be given by the name of a R color, the name of a CSS
color, e.g. \code{"transparent"} or \code{"fuchsia"}, an HEX code like
\code{"#ff009a"}, a RGB code like \code{"rgb(255,100,39)"}, or a HSL code
like \code{"hsl(360,11,255)"}.
}
\examples{
# a line chart with numeric x-axis ####

set.seed(666)
dat <- data.frame(
  x = 1:10,
  y1 = rnorm(10),
  y2 = rnorm(10)
)

amLineChart(
  data = dat,
  width = "700px",
  xValue = "x",
  yValues = c("y1", "y2"),
  yValueNames = list(y1 = "Sample 1", y2 = "Sample 2"),
  draggable = list(y1 = TRUE, y2 = FALSE),
  backgroundColor = "#30303d",
  lineStyle = list(
    color = list(y1 = "yellow", y2 = "orangered"),
    width = 4
  ),
  chartTitle = list(text = "Gaussian samples", color = "whitesmoke"),
  xAxis = list(title = list(text = "Observation",
                            fontSize = 21,
                            color = "silver"),
               labels = list(color = "whitesmoke",
                             fontSize = 17)),
  yAxis = list(title = list(text = "Value",
                            fontSize = 21,
                            color = "silver"),
               labels = list(color = "whitesmoke",
                             fontSize = 14)),
  minY = -3, maxY = 3,
  valueFormatter = "#.#",
  caption = list(text = "[font-style:italic]rAmCharts4[/]",
                 color = "yellow"),
  gridLines = list(color = "whitesmoke",
                   opacity = 0.4,
                   width = 1),
  theme = "dark")


# line chart with date x-axis ####

library(lubridate)

set.seed(666)
dat <- data.frame(
  date = ymd(180101) + months(0:11),
  visits = rpois(12, 20)
)

amLineChart(
  data = dat,
  width = "700px",
  xValue = "date",
  yValues = "visits",
  draggable = TRUE,
  chartTitle = "Number of visits",
  xAxis = "Date",
  yAxis = "Visits",
  minY = 0, maxY = 35,
  backgroundColor = "whitesmoke",
  tooltip = "[bold][font-style:italic]{dateX}[/]\nvisits: {valueY}[/]",
  valueFormatter = "#",
  caption = list(text = "Year 2018"),
  theme = "material")
}
