% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amLineChart.R
\name{amLineChart}
\alias{amLineChart}
\title{Create a HTML widget displaying a line chart}
\usage{
amLineChart(
  data,
  data2 = NULL,
  xValue,
  yValues,
  yValueNames = NULL,
  xLimits = NULL,
  yLimits = NULL,
  expandX = 0,
  expandY = 5,
  valueFormatter = "#.",
  trend = FALSE,
  chartTitle = NULL,
  theme = NULL,
  draggable = FALSE,
  tooltip = NULL,
  lineStyle = NULL,
  backgroundColor = NULL,
  xAxis = NULL,
  yAxis = NULL,
  scrollbarX = FALSE,
  scrollbarY = FALSE,
  gridLines = NULL,
  legend = NULL,
  caption = NULL,
  button = NULL,
  width = NULL,
  height = NULL,
  chartId = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{a dataframe}

\item{data2}{\code{NULL} or a dataframe used to update the data with the
button; its column names must include the column names of \code{data}
given in \code{values} and it must have the same number of rows as
\code{data}}

\item{xValue}{name of the column of \code{data} to be used on the
x-axis}

\item{yValues}{name(s) of the column(s) of \code{data} to be used on the
y-axis}

\item{yValueNames}{names of the variables on the y-axis,
to appear in the legend;
\code{NULL} to use \code{yValues} as names, otherwise a named list of the
form \code{list(yvalue1 = "ValueName1", yvalue2 = "ValueName2", ...)} where
\code{yvalue1}, \code{yvalue2}, ... are the column names given in
\code{yValues} and \code{"ValueName1"}, \code{"ValueName2"}, ... are the
desired names to appear in the legend}

\item{xLimits}{range of the x-axis, a vector of two values specifying
the left and the right limits of the x-axis; \code{NULL} for default values}

\item{yLimits}{range of the y-axis, a vector of two values specifying
the lower and the upper limits of the y-axis; \code{NULL} for default values}

\item{expandX}{if \code{xLimits = NULL}, a percentage of the range of the
x-axis used to expand this range}

\item{expandY}{if \code{yLimits = NULL}, a percentage of the range of the
y-axis used to expand this range}

\item{valueFormatter}{a number formatter for XXXX; see
\url{https://www.amcharts.com/docs/v4/concepts/formatters/formatting-numbers/}}

\item{trend}{option to request trend lines and to set their settings;
\code{FALSE} for no trend line, otherwise a named list of the form
\code{list(yvalue1 = trend1, yvalue2 = trend2, ...)} where
\code{trend1}, \code{trend2}, ... are lists with the following fields:
\describe{
  \item{\code{method}}{
    the modelling method, can be \code{"lm"}, \code{"lm.js"}
    or \code{"loess"}; \code{"lm.js"} performs a polynomial regression in
    JavaScript, its advantage is that the fitted regression line is
    updated when the points of the line are dragged
  }
  \item{\code{formula}}{
    a formula passed to the \code{lm} function if \code{method = "lm"}; the
    lefthandside of this formula must always be \code{y}, and its
    righthandside must be a symbolic expression depending on \code{x} only,
    e.g. \code{y ~ x}, \code{y ~ x + I(x^2)}, \code{y ~ poly(x,2)}
  }
  \item{\code{order}}{
    the order of the polynomial regression when \code{method = "lm.js"}
  }
  \item{\code{method.args}}{
    a list of additional arguments passed on to the modelling function
    defined by \code{method} when \code{method = "loess"}, e.g.
    \code{list(span = 0.3)}
  }
  \item{\code{style}}{
    a list of settings for the trend line created with \code{\link{amLine}}
  }
}
it is also possible to request the same kind of trend lines for all series
given by the \code{yValues} argument, by passing a list of the
form \code{list("_all" = trendconfig)}, e.g.
\code{list("_all" = list(method = "lm", formula = y ~ 0+x, style = amLine()))}}

\item{chartTitle}{chart title, \code{NULL}, character, or list of settings}

\item{theme}{theme, \code{NULL} or one of \code{"dataviz"},
\code{"material"}, \code{"kelly"}, \code{"dark"}, \code{"moonrisekingdom"},
\code{"frozen"}, \code{"spiritedaway"}, \code{"patterns"},
\code{"microchart"}}

\item{draggable}{\code{TRUE}/\code{FALSE} to enable/disable dragging of
all lines, otherwise a named list of the form
\code{list(yvalue1 = TRUE, yvalue2 = FALSE, ...)} to enable/disable the
dragging for each bar corresponding to a column given in \code{yValues}}

\item{tooltip}{tooltip settings given as a list, or just a string for the
\code{text} field, or \code{FALSE} for no tooltip, or \code{NULL} for the
default tooltip; the \code{text} field must be a formatted string:
\url{https://www.amcharts.com/docs/v4/concepts/formatters/formatting-strings/}}

\item{lineStyle}{settings of the lines; \code{NULL} for default,
otherwise a named list of the form
\code{list(yvalue1 = settings1, yvalue2 = settings2, ...)} where
\code{settings1}, \code{settings2}, ... are lists created with
\code{\link{amLine}}; this can also be a
single list of settings that will be applied to each line}

\item{backgroundColor}{a color for the chart background}

\item{xAxis}{settings of the x-axis given as a list, or just a string
for the axis title}

\item{yAxis}{settings of the y-axis given as a list, or just a string
for the axis title}

\item{scrollbarX}{logical, whether to add a scrollbar for the category axis}

\item{scrollbarY}{logical, whether to add a scrollbar for the value axis}

\item{gridLines}{settings of the grid lines}

\item{legend}{logical, whether to display the legend}

\item{caption}{settings of the caption, or \code{NULL} for no caption}

\item{button}{\code{NULL} for the default, \code{FALSE} for no button,
a single character string giving the button label,
or settings of the button given as
a list with these fields: \code{text} for the button label, \code{color} for
the label color, \code{fill} for the button color, and \code{position}
for the button position as a percentage (\code{0} for bottom,
\code{1} for top); this button is used to replace the current data
with \code{data2}}

\item{width}{the width of the chart, e.g. \code{"600px"} or \code{"80\%"};
ignored if the chart is displayed in Shiny, in which case the width is
given in \code{\link{amChart4Output}}}

\item{height}{the height of the chart, e.g. \code{"400px"};
ignored if the chart is displayed in Shiny, in which case the width is
given in \code{\link{amChart4Output}}}

\item{chartId}{a HTML id for the chart}

\item{elementId}{a HTML id for the container of the chart; ignored if the
chart is displayed in Shiny, in which case the id is given by the Shiny id}
}
\description{
Create a HTML widget displaying a line chart.
}
\note{
A color can be given by the name of a R color, the name of a CSS
color, e.g. \code{"transparent"} or \code{"fuchsia"}, an HEX code like
\code{"#ff009a"}, a RGB code like \code{"rgb(255,100,39)"}, or a HSL code
like \code{"hsl(360,11,255)"}.
}
\examples{
# a line chart with numeric x-axis ####

set.seed(666)
dat <- data.frame(
  x = 1:10,
  y1 = rnorm(10),
  y2 = rnorm(10)
)

amLineChart(
  data = dat,
  width = "700px",
  xValue = "x",
  yValues = c("y1", "y2"),
  yValueNames = list(y1 = "Sample 1", y2 = "Sample 2"),
  trend = list(
    y1 = list(
      method = "lm",
      formula = y ~ poly(x,2),
      style = amLine(color = "lightyellow", dash = "3,2")
    ),
    y2 = list(
      method = "loess",
      style = amLine(color = "palevioletred", dash = "3,2")
    )
  ),
  draggable = list(y1 = TRUE, y2 = FALSE),
  backgroundColor = "#30303d",
  lineStyle = list(
    y1 = amLine(color = "yellow", width = 4),
    y2 = amLine(color = "orangered", width = 4)
  ),
  chartTitle = list(text = "Gaussian samples", color = "whitesmoke"),
  xAxis = list(title = list(text = "Observation",
                            fontSize = 21,
                            color = "silver"),
               labels = list(color = "whitesmoke",
                             fontSize = 17)),
  yAxis = list(title = list(text = "Value",
                            fontSize = 21,
                            color = "silver"),
               labels = list(color = "whitesmoke",
                             fontSize = 14)),
  yLimits = c(-3, 3),
  valueFormatter = "#.#",
  caption = list(text = "[font-style:italic]rAmCharts4[/]",
                 color = "yellow"),
  gridLines = list(color = "whitesmoke",
                   opacity = 0.4,
                   width = 1),
  theme = "dark")


# line chart with date x-axis ####

library(lubridate)

set.seed(666)
dat <- data.frame(
  date = ymd(180101) + months(0:11),
  visits = rpois(12, 20)
)

amLineChart(
  data = dat,
  width = "700px",
  xValue = "date",
  yValues = "visits",
  draggable = TRUE,
  chartTitle = "Number of visits",
  xAxis = "Date",
  yAxis = "Visits",
  yLimits = c(0, 35),
  backgroundColor = "whitesmoke",
  tooltip = "[bold][font-style:italic]{dateX}[/]\nvisits: {valueY}[/]",
  valueFormatter = "#",
  caption = list(text = "Year 2018"),
  theme = "material")


# smoothed lines ####

x <- seq(-4, 4, length.out = 100)
dat <- data.frame(
  x = x,
  Gauss = dnorm(x),
  Cauchy = dcauchy(x)
)

amLineChart(
  data = dat,
  width = "700px",
  xValue = "x",
  yValues = c("Gauss", "Cauchy"),
  yValueNames = list(
    Gauss = "Standard normal distribution",
    Cauchy = "Cauchy distribution"
  ),
  draggable = FALSE,
  tooltip = FALSE,
  lineStyle = amLine(
    width = 4,
    tensionX = 0.8,
    tensionY = 0.8
  ),
  xAxis = list(title = list(text = "x",
                            fontSize = 21,
                            color = "navyblue"),
               labels = list(color = "midnightblue",
                             fontSize = 17)),
  yAxis = list(title = list(text = "density",
                            fontSize = 21,
                            color = "navyblue"),
               labels = FALSE),
  theme = "dataviz")
}
