% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{rAmCharts4-adapters}
\alias{rAmCharts4-adapters}
\alias{amColorAdapterFromVector}
\alias{amColorAdapterFromCuts}
\title{Adapters}
\usage{
amColorAdapterFromVector(colors)

amColorAdapterFromCuts(cuts, colors, axis)
}
\arguments{
\item{colors}{vector of colors}

\item{cuts}{a vector of cut points (sorted increasingly)}

\item{axis}{\code{"X"} or \code{"Y"}, the axis to which the current value
refers to}
}
\description{
Adapters allow to have finer control of settings such as
  the colors of the columns of a bar chart or the colors of the points
  of a scatter chart.
}
\examples{
# bar chart with individual colors ####

dat <- data.frame(
  country = c("USA", "China", "Japan", "Germany", "UK", "France"),
  visits = c(3025, 1882, 1809, 1322, 1122, 1114)
)

amBarChart(
  data = dat,
  width = "600px",
  category = "country", values = "visits",
  showValues = FALSE,
  tooltip = FALSE,
  columnStyle = amColumn(
    color = amColorAdapterFromVector(hcl.colors(6, "Viridis")),
    opacity = 0.7,
    strokeColor = amColorAdapterFromVector(hcl.colors(6, "Cividis")),
    strokeWidth = 4
  ),
  bullets = amCircle(
    color = amColorAdapterFromVector(hcl.colors(6, "Viridis")),
    opacity = 1,
    strokeColor = amColorAdapterFromVector(hcl.colors(6, "Cividis")),
    strokeWidth = 4,
    radius = 12
  ),
  alwaysShowBullets = TRUE,
  chartTitle =
    list(text = "Visits per country", fontSize = 22, color = "orangered"),
  backgroundColor = "rgb(164,167,174)",
  xAxis = list(title = list(text = "Country", color = "maroon")),
  yAxis = list(
    title = list(text = "Visits", color = "maroon"),
    gridLines = amLine(color = "white", width = 1, dash = "3,3")
  ),
  yLimits = c(0, 4000),
  valueFormatter = "#,###.",
  caption = list(text = "Year 2018", color = "red")
)


# usage example of amColorAdapterFromCuts ####

set.seed(314159)
dat <- data.frame(
  x = rnorm(200),
  y = rnorm(200)
)

amScatterChart(
  data = dat,
  width = "500px", height = "500px",
  xValue = "x", yValues = "y",
  xLimits = c(-3,3), yLimits = c(-3,3),
  draggable = FALSE,
  backgroundColor = "#30303d",
  pointsStyle = amCircle(
    color = amColorAdapterFromCuts(
      cuts = c(-2, -1, 1, 2),
      colors = c("red", "green", "blue", "green", "red"),
      axis = "Y" # try to change to "X"
    ),
    opacity = 0.5,
    strokeColor = amColorAdapterFromCuts(
      cuts = c(-2, -1, 1, 2),
      colors = c("darkred", "darkgreen", "darkblue", "darkgreen", "darkred"),
      axis = "Y" # try to change to "X"
    )
  ),
  xAxis = list(
    breaks = amAxisBreaks(seq(-3, 3, by=1)),
    gridLines = amLine(opacity = 0.3, width = 1)
  ),
  yAxis = list(
    breaks = amAxisBreaks(seq(-3, 3, by=1)),
    gridLines = amLine(opacity = 0.3, width = 1)
  ),
  tooltip = FALSE,
  caption = list(text = "[font-style:italic]rAmCharts4[/]",
                 color = "yellow"),
  theme = "dark")
}
