% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullets.R
\name{rAmCharts4-shapes}
\alias{rAmCharts4-shapes}
\alias{amTriangle}
\alias{amCircle}
\alias{amRectangle}
\title{Bullets}
\usage{
amTriangle(
  color = NULL,
  width = 10,
  height = 10,
  strokeColor = NULL,
  strokeWidth = 2,
  direction = "top",
  rotation = 0
)

amCircle(color = NULL, radius = 6, strokeColor = NULL, strokeWidth = 2)

amRectangle(
  color = NULL,
  width = 10,
  height = 10,
  strokeColor = NULL,
  strokeWidth = 2,
  rotation = 0,
  cornerRadius = 3
)
}
\arguments{
\item{color}{bullet color}

\item{width}{bullet width}

\item{height}{bullet height}

\item{strokeColor}{stroke color of the bullet}

\item{strokeWidth}{stroke width of the bullet}

\item{direction}{triangle direction}

\item{rotation}{rotation angle}

\item{radius}{circle radius}

\item{cornerRadius}{radius of the rectangle corners}
}
\value{
A list of settings for the bullets.
}
\description{
Create a list of settings for bullets,
their shape and their style.
}
