% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{rAmCharts4-shiny}
\alias{rAmCharts4-shiny}
\alias{amChart4Output}
\alias{renderAmChart4}
\title{Shiny bindings for using \code{rAmCharts4} in Shiny}
\usage{
amChart4Output(outputId, width = "100\%", height = "400px")

renderAmChart4(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended}

\item{expr}{an expression that generates a chart with
\code{\link{amBarChart}}, \code{\link{amHorizontalBarChart}} or
\code{\link{amLineChart}}}

\item{env}{the environment in which to evaluate \code{expr}}

\item{quoted}{whether \code{expr} is a quoted expression}
}
\description{
Output and render functions for using
the \code{rAmCharts4} widgets
within Shiny applications and interactive Rmd documents.
}
\examples{
if(interactive()) {
library(rAmCharts4)
library(shiny)
library(lubridate)

ui <- fluidPage(
  br(),
  fluidRow(
    column(
      width = 8,
      amChart4Output("linechart", height = "500px")
    ),
    column(
      width = 4,
      verbatimTextOutput("chartData"),
      verbatimTextOutput("chartChange")
    )
  )
)

server <- function(input, output){

  set.seed(666)
  dat <- data.frame(
    date = ymd(180101) + months(0:11),
    visits = rpois(12, 20),
    x = 1:12
  )

  output[["linechart"]] <- renderAmChart4({
    amLineChart(
      data = dat,
      data2 = dat,
      xValue = "date",
      yValues = "visits",
      draggable = TRUE,
      chartTitle = list(text = "Number of visits", color = "crimson"),
      xAxis = list(title = "Date", labels = list(rotation = -45)),
      yAxis = "Visits",
      minY = 0, maxY = 35,
      backgroundColor = "whitesmoke",
      tooltip = "[bold][font-style:italic]{dateX}[/]\nvisits: {valueY}[/]",
      valueFormatter = "#",
      caption = list(
        text = "[bold font-size:22]Year 2018[/]",
        color = "fuchsia"
      ),
      button = list(
        text = "Reset data",
        color = "black",
        fill = "seashell",
        position = 0.9
      ),
      theme = "dataviz")
  })


  output[["chartData"]] <- renderPrint({
    input[["linechart"]]
  })

  output[["chartChange"]] <- renderPrint({
    input[["linechart_change"]]
  })

}

shinyApp(ui, server)
}
}
