% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyIcon.R
\name{tinyIcon}
\alias{tinyIcon}
\alias{tinyIcons}
\alias{shinyAppTinyIcons}
\title{Icons}
\usage{
tinyIcon(icon, backgroundColor = NULL)

tinyIcons()

shinyAppTinyIcons()
}
\arguments{
\item{icon}{name of an icon; \code{tinyIcons()} returns the list of
available icons, and \code{shinyAppTinyIcons()} runs a Shiny app which
displays the available icons}

\item{backgroundColor}{background color of the icon
(possibly \code{"transparent"})}
}
\value{
A base64 string that can be used in the \code{href} argument of
  \code{\link{amImage}}.
}
\description{
Icons for usage in \code{\link{amImage}}.
}
\note{
A color can be given by the name of a R color, the name of a CSS
  color, e.g. \code{"transparent"} or \code{"fuchsia"}, an HEX code like
  \code{"#ff009a"}, a RGB code like \code{"rgb(255,100,39)"}, or a HSL code
  like \code{"hsl(360,11,255)"}.
}
